feat: add interactive intraday chart to asset detail page (#435)
Add Live/Historical mode toggle to the asset detail header. "Live" shows
an interactive intraday chart (zoom, scroll, magnet crosshair, session
colors) powered by SSE data; "Historical" preserves the existing
candlestick chart with period selector.

Extract chart rendering from IntradayMountainChart into a shared
IntradayChart component with compact/interactive modes, reused by both
the group page LiveDayView and the asset detail page.
